Fix packaging: move files and update rockspec
Rockspec's install has a number of odd behaviors... for example, export
filename and path must match source filename and path. Annoying (why not
do implicit conversion if you're stuck with the source files in the same
layout?). This also means we have to install the rock to use it, or
paths will not match. Not sure if there is any way around that.